refactor(ui): extract SurveyView trailing helpers
Move renderSampleData and calculateMetricRange out of SurveyView.tsx
into surveyViewHelpers.tsx (~168 lines extracted). SurveyView.tsx
still exceeds the 800-line red-flag budget at 1650 lines and needs
further work to break up the main component body; the helpers split
is a first step that isolates the two pure-function helpers that
don't depend on component state.

Pure relocation, no behavior change.
